portalservice/requirements.txt
apistar==0.7.2
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dateutils==0.6.12
Django==4.0.4
django-appconf==1.0.5
django-cors-headers==3.12.0
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-nine==0.2.5
django-opensearch==0.20.0
django-opensearch-dsl==0.2.0
django-request-logging==0.7.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
opensearch-dsl==1.0.0
opensearch-py==1.1.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typesystem==0.4.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
